+++ CHANGES.rst
@@ -3,3 +3,4 @@
 
 * base dependencies + extras (halotools, h5py); install all dependencies via pip nbodykit[extras]
 * meta-data calculations in FKPCatalog now account for Source selection properly
+* support for numpy int/float meta-data in JSON output files